Modified order of functions and function names
Made function names more consistent and clear.  Changed the old way
of entire process being an arrow function to the game being it's own
function that is called by the arrow function.  Made a function to
reset that player score that is called by the old function that
determines the winner.
Reorganized the order of functions to match the order called by the
playGame() function. Modified so capitalizations are consistent and
names of html variables are consistent.